+++ robot_descriptions/luna_mj_description.py
@@ -1,0 +1,19 @@
+#!/usr/bin/env python3
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Luna MJCF description."""
+
+from os import getenv as _getenv
+from os import path as _path
+
+from ._cache import clone_to_cache as _clone_to_cache
+
+REPOSITORY_PATH: str = _clone_to_cache(
+    "luna_description",
+    commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
+)
+
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "HU_L04_description")
+
+MJCF_PATH: str = _path.join(PACKAGE_PATH, "xml", "HU_L04_01.xml")